requirements.txt
# API Framework
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
jinja2==3.1.3

# ML / NLP
transformers==4.38.2
torch==2.2.1          # CPU-only: pip install torch --index-url https://download.pytorch.org/whl/cpu
sentencepiece==0.1.99

# TTS engines
gTTS==2.5.1
pyttsx3==2.90

# Audio processing
pydub==0.25.1
# Note: pydub needs ffmpeg installed at OS level (not pip)

# Utilities
python-dotenv==1.0.1
diskcache==5.6.3

# Optional: ElevenLabs
# elevenlabs==0.2.27
